storyvord/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
azure-core==1.30.2
azure-storage-blob==12.20.0
billiard==4.2.0
black==24.4.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
daphne==4.1.2
decorator==5.1.1
defusedxml==0.8.0rc2
dill==0.3.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-model-utils==4.3.1
django-rest-swagger==2.2.0
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.1
executing==2.0.1
geonamescache==2.0.0
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.0
grpcio-status==1.65.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
isodate==0.6.1
isort==5.13.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack==1.0.8
mypy-extensions==1.0.0
numpy==2.0.0
oauthlib==3.2.2
openai==1.40.0
openapi-codec==1.3.2
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pika==1.3.2
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.5
pyOpenSSL==24.1.0
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==7.0.3
django-allauth>=0.43.0
django-unfold==0.42.0